working stage with messy (beautiful) UI
This is a working app with a nice UI that is not efficient according to some people. Therefore I need to make it boring. Note: not all the sample size calculation methods are working currently and not all of them are added to the app. flowchart still needs to be updated.